Another major client refactoring.
All client session state is now stored in a client_session struct,
rather than in two unconnected watcher_pair structs, as before. For
now, there's no functional difference, but storing all state in a
single struct will make it much easier to make session-wide changes,
such as cleanly shutting down a connection when we receive a premature
EOF from the server, or when a session-fatal error occurs.

This approach requires separate callback stubs for each of the 4 libev
ev_io watchers (stdin, stdout, and the server read/write watchers),
and for each of the 3 timeout callbacks (stdin, server read/write).
It's more code than the old way required, but the bulk of the callback
callback code is still shared by all in read_cb, write_cb, and
echo_proto_timeout_cb. The separate callback stubs are relatively
simple and short.

Storing session state in a per-session struct is complete overkill for
a simple command-line echo client (we could simply store everything in
global variables), but once all of the exit(3) calls are factored out
of the ev_io callbacks and timer callbacks, it'll be easy to re-use
the code in this example client in more complicated programs, such as
event-driven GUI apps, where the app may live beyond a single session.